requirements.txt
aenum==3.1.12
allure-pytest==2.13.2
allure-python-commons==2.13.2
anyio==3.6.2
assertpy==1.1
astroid==2.15.4
async-generator==1.10
attrs==23.1.0
azure-common==1.1.28
azure-core==1.26.4
azure-identity==1.12.0
azure-keyvault-secrets==4.7.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.2.1
azure-mgmt-resource==23.0.0
azure-storage-blob==12.16.0
azurify @ file:///Users/deniz/Dev/azurify
beautifulsoup4==4.12.2
black==23.1.0
bleach==6.0.0
blinker==1.6.2
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
ConfigArgParse==1.5.3
cryptography==40.0.2
curlify==2.2.1
decorator==5.1.1
delayed-assert==0.3.6
dill==0.3.6
docutils==0.19
exceptiongroup==1.1.1
execnet==1.9.0
exrex==0.11.0
Faker==18.6.1
fast-map==0.0.7
fastapi==0.95.1
fastapi-utils==0.2.1
flake8==6.0.0
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==2.0.9
greenlet==1.1.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
influxdb-client==1.36.1
iniconfig==2.0.0
isodate==0.6.1
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
jsonpath-ng==1.5.3
keyring==23.13.1
lazy-object-proxy==1.9.0
locust==2.12.1
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.1.0
msal==1.22.0
msal-extensions==1.0.0
msgpack==1.0.5
mypy-extensions==1.0.0
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==1.5.0
pathspec==0.11.1
pkginfo==1.9.6
platformdirs==3.5.0
pluggy==1.0.0
ply==3.11
portalocker==2.7.0
prometheus-client==0.16.0
psutil==5.9.5
psycogreen==1.0.2
psycopg2-binary==2.9.6
PyByteBuffer==1.0.5
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.6.0
pylint==2.17.0
PySocks==1.7.1
pytest==7.3.1
pytest-json-report==1.5.0
pytest-metadata==2.0.4
pytest-order==1.1.0
pytest-reportportal==5.1.5
pytest-rerunfailures==11.1.2
pytest-xdist==3.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.0.2
reactivex==4.0.4
readme-renderer==37.3
reportportal-client==5.2.5
requests==2.29.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.3.5
roundrobin==0.0.4
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
starlette==0.26.1
StrEnum==0.4.10
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
twine==4.0.2
typing_extensions==4.5.0
tzdata==2023.3
unitest==1.4.34
urllib3==1.26.15
uvicorn==0.22.0
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.3.3
wrapt==1.15.0
wsproto==1.2.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
